Enhancements for UE4 Editor
Warning: This may not compile under UE 4.17 or earlier (could be fixable later I guess).

Added support for most of the V3D command events provided by 3dconnexion - but requires the user to have a proper configuration for the 3D mouse.
The V3D commands will execute the equivalent viewport commands in the UE4 Editor - such as Front, Left, Fit, etc. I hope it is okay that I call the commands directly from within the plugin. I am not sure if this is the proper way to do it.